requirements.txt

fastapi==0.115.11
uvicorn==0.29.0
python-dotenv==1.0.1
fastmcp
anthropic
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sse-starlette
asyncio>=3.4.3
browser-use>=0.1.40
click>=8.1.8
httpx>=0.28.1
langchain-openai>=0.3.1
mcp>=1.3.0
pydantic>=2.10.6
anyio
playwright>=1.50.0